requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
aiosqlite==0.19.0
free-proxy==1.1.1
APScheduler==3.10.4
pytz==2023.3
python-dotenv==1.0.0
numpy>=1.24.0
Pillow>=10.0.0
lxml>=4.9.0
telegram==0.0.1